Rename new module
